+++ app/src/js/config.js
@@ -1,12 +1,13 @@
-window.siyuan = {
-    config: {
-        appearance: { mode: 0, codeBlockThemeDark: "atom-one-dark", codeBlockThemeLight: "atom-one-light" },
-        editor: {
-            codeLineWrap: true,
-            codeLigatures: true,
-            codeSyntaxHighlightLineNum: true,
-            plantUMLServePath: "https://www.plantuml.com/plantuml/svg/~1",
-        },
-    },
-    languages: { copy: "复制" },
-};
+/* 覆盖 *.config.toml 配置文件中定义的字段 */
+// window.siyuan = {
+//     config: {
+//         appearance: { mode: 0, codeBlockThemeDark: "atom-one-dark", codeBlockThemeLight: "atom-one-light" },
+//         editor: {
+//             codeLineWrap: true,
+//             codeLigatures: true,
+//             codeSyntaxHighlightLineNum: true,
+//             plantUMLServePath: "https://www.plantuml.com/plantuml/svg/~1",
+//         },
+//     },
+//     languages: { copy: "复制" },
+// };